// src/types/Training.ts
export interface TrainingData {
  Name: string;
  "Last name": string;
  WWID: string;
  Region: string;
  Country: string;
  "Operating Group": string;
  Sector: string;
  Type: string;
  Results: string;
  Realized: string;
  "Expiration Date": string;
  Status: string;
}
export interface TrainingBTWData {
  Name: string;
  "Last name": string;
  WWID: string;
  Region: string;
  Country: string;
  "Operating Group": string;
  Sector: string;
  Type: string;
  Results: string;
  Realized: string;
  "Expiration Date": string;
  Status: string;
}